1) BUG FIX- bug description- In the quiz page, whenever i would refresh the page would error: quiz not defined.
bug cause-
The page rendering was dependent on the data from the server and since it is loaded in the quizList component if i entered the quiz
page straight away I did not have the data.

bug fix-
First I fetch the data inside the quiz component as well. However it was still not enough since the fetching process async and
therefore the page could still not render immediately.
The fix for this was to render a spinner untill the data was loaded.

2) Changed the serve route from '/' to the catch all '*', thus every request will serve the app.